Security applied, changes verifications controllers, authentication, authorization

+++ src/main/java/Me/ChallengeJava/controllers/CourseController.java
@@ -30,20 +30,32 @@ public class CourseController {
     // Create a new course
     @PostMapping("/api/courses")
     public ResponseEntity<Object> createCourse(@RequestBody CourseApplicationDTO courseApplicationDTO){
+        // Validation name
         if(courseApplicationDTO.getName().isBlank()){
             return new ResponseEntity<>("Name can't be on blank", HttpStatus.FORBIDDEN);
         }
+        // Validation description
         if(courseApplicationDTO.getDescription().isBlank()){
             return new ResponseEntity<>("Description can't be on blank", HttpStatus.FORBIDDEN);
         }
+        // Validation shifts
         if(courseApplicationDTO.getShifts().isEmpty()){
             return new ResponseEntity<>("Shifts can't be empty", HttpStatus.FORBIDDEN);
+        }
+        // Validation imageUrl
+        if(courseApplicationDTO.getImageUrl().isBlank()){
+            return new ResponseEntity<>("Image Url can't be on blank", HttpStatus.FORBIDDEN);
+        }
+        if(courseApplicationDTO.getImageUrl().endsWith(".png") &&
+                courseApplicationDTO.getImageUrl().endsWith(".jpeg")){
+            return new ResponseEntity<>("Format is not permitted", HttpStatus.FORBIDDEN);
         }
         // Create new course
         Course newCourse = new Course(
                 courseApplicationDTO.getName(), // Name
                 courseApplicationDTO.getDescription(), // Description
-                courseApplicationDTO.getShifts().stream().collect(toList())); // Shifts
+                courseApplicationDTO.getShifts().stream().collect(toList()),
+                courseApplicationDTO.getImageUrl()); // Shifts
         // Save new course
         courseRepository.save(newCourse);
 
@@ -78,6 +90,7 @@ public class CourseController {
             courseToModify.setName(courseApplicationDTO.getName()); // Name
             courseToModify.setDescription(courseApplicationDTO.getDescription()); // Description
             courseToModify.setShifts(courseApplicationDTO.getShifts().stream().collect(toList())); // Shifts
+            courseToModify.setImageUrl(courseToModify.getImageUrl()); // Image URL
             // Save course
             courseRepository.save(courseToModify);
 
@@ -86,9 +99,15 @@ public class CourseController {
         return new ResponseEntity<>("Course doesn't exist", HttpStatus.FORBIDDEN);
     }
     @DeleteMapping("/api/courses/{id}")
-    public void deleteCourse(@PathVariable Long id){
-        Course courseToDelete = courseRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Course not found for this id ::" + id));
+    public ResponseEntity<?> deleteCourse(@PathVariable Long id){
+        // Get course
+        Course courseToDelete = courseRepository.findById(id).orElse(null);
+        // Verification course
+        if(courseToDelete == null){
+            return new ResponseEntity<>("This course doesn't exist", HttpStatus.FORBIDDEN);
+        }
         // Delete user
         courseRepository.delete(courseToDelete);
+        return new ResponseEntity<>("Course deleted successfully", HttpStatus.OK);
     }
 }
